fix(core): Improve SynthesizerHub._drain_all_synthesizers
Improve the way how it is checked whether the running Synthesizer has
finished.

As of commit 3ec0565, SynthesizerHub checked whether the queue of a
Synthesizer is empty to determine if the Synthesizer has finished,
but this may miss those synthesizers which are synthesizing and not
serving their buffer in the queue.